// src/enums/piece-id.enum.ts
export enum PieceId {
  BLACK_BISHOP_1 = 'black_bishop_1',
  BLACK_BISHOP_2 = 'black_bishop_2',
  BLACK_KING = 'black_king',
  BLACK_KNIGHT_1 = 'black_knight_1',
  BLACK_KNIGHT_2 = 'black_knight_2',
  BLACK_PAWN_1 = 'black_pawn_1',
  BLACK_PAWN_2 = 'black_pawn_2',
  BLACK_PAWN_3 = 'black_pawn_3',
  BLACK_PAWN_4 = 'black_pawn_4',
  BLACK_PAWN_5 = 'black_pawn_5',
  BLACK_PAWN_6 = 'black_pawn_6',
  BLACK_PAWN_7 = 'black_pawn_7',
  BLACK_PAWN_8 = 'black_pawn_8',
  BLACK_QUEEN = 'black_queen',
  BLACK_ROOK_1 = 'black_rook_1',
  BLACK_ROOK_2 = 'black_rook_2',
  WHITE_BISHOP_1 = 'white_bishop_1',
  WHITE_BISHOP_2 = 'white_bishop_2',
  WHITE_KING = 'white_king',
  WHITE_KNIGHT_1 = 'white_knight_1',
  WHITE_KNIGHT_2 = 'white_knight_2',
  WHITE_PAWN_1 = 'white_pawn_1',
  WHITE_PAWN_2 = 'white_pawn_2',
  WHITE_PAWN_3 = 'white_pawn_3',
  WHITE_PAWN_4 = 'white_pawn_4',
  WHITE_PAWN_5 = 'white_pawn_5',
  WHITE_PAWN_6 = 'white_pawn_6',
  WHITE_PAWN_7 = 'white_pawn_7',
  WHITE_PAWN_8 = 'white_pawn_8',
  WHITE_QUEEN = 'white_queen',
  WHITE_ROOK_1 = 'white_rook_1',
  WHITE_ROOK_2 = 'white_rook_2',
}